refactor: move createInvite to inviteservice so we can use it in seed
